move MigrateCost to MoveCost (#1585)
