+++ common/src/main/java/org/astraea/common/cost/HasMoveCost.java
@@ -40,10 +40,11 @@ public interface HasMoveCost extends CostFunction {
 
       @Override
       public MoveCost moveCost(ClusterInfo before, ClusterInfo after, ClusterBean clusterBean) {
-        var overflow =
+        var moveCosts =
             hasMoveCosts.stream()
                 .map(c -> c.moveCost(before, after, clusterBean))
-                .anyMatch(MoveCost::overflow);
+                .collect(Collectors.toSet());
+        var overflow = moveCosts.stream().anyMatch(MoveCost::overflow);
         return () -> overflow;
       }
