+++ common/src/main/java/org/astraea/common/cost/MoveCost.java
@@ -17,7 +17,6 @@
 package org.astraea.common.cost;
 
 /** Return type of cost function, `HasMoveCost`. It returns the score of migrate plan. */
-@FunctionalInterface
 public interface MoveCost {
 
   MoveCost EMPTY = () -> false;
