+++ common/src/main/java/org/astraea/common/cost/ReplicaLeaderSizeCost.java
@@ -16,6 +16,8 @@
  */
 package org.astraea.common.cost;
 
+import static org.astraea.common.cost.MigrationCost.changedRecordSizeOverflow;
+
 import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -33,10 +35,10 @@ import org.astraea.common.metrics.collector.MetricSensor;
  */
 public class ReplicaLeaderSizeCost
     implements HasMoveCost, HasBrokerCost, HasClusterCost, HasPartitionCost {
-
   private final Configuration config;
 
   public static final String COST_LIMIT_KEY = "max.migrated.leader.size";
+  public static final String MOVED_LEADER_SIZE = "moved leader size (bytes)";
 
   public ReplicaLeaderSizeCost() {
     this.config = Configuration.of(Map.of());
@@ -61,8 +63,7 @@ public class ReplicaLeaderSizeCost
     var maxMigratedLeaderSize =
         config.string(COST_LIMIT_KEY).map(DataSize::of).map(DataSize::bytes).orElse(Long.MAX_VALUE);
     var overflow =
-        ClusterInfo.changedRecordSizeOverflow(
-            before, after, Replica::isLeader, maxMigratedLeaderSize);
+        changedRecordSizeOverflow(before, after, Replica::isLeader, maxMigratedLeaderSize);
     return () -> overflow;
   }
