+++ common/src/test/java/org/astraea/common/admin/ClusterInfoTest.java
@@ -24,8 +24,6 @@ import java.util.concurrent.ThreadLocalRandom;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import org.astraea.common.Utils;
-import org.astraea.common.cost.ReplicaLeaderCost;
-import org.astraea.common.cost.ReplicaNumberCost;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
@@ -111,392 +109,5 @@ public class ClusterInfoTest {
         Exception.class, () -> cluster.replicas(TopicPartition.of("t", 0)).add(replica));
     Assertions.assertThrows(
         Exception.class, () -> cluster.replicas(TopicPartitionReplica.of("t", 0, 10)).add(replica));
-  }
-
-  @Test
-  void testChangedReplicaLeaderNumber() {
-    var costFunction = new ReplicaLeaderCost();
-    var before =
-        List.of(
-            Replica.builder()
-                .topic("topic1")
-                .partition(0)
-                .nodeInfo(NodeInfo.of(0, "broker0", 1111))
-                .lag(-1)
-                .size(-1)
-                .isLeader(true)
-                .isSync(true)
-                .isFuture(false)
-                .isOffline(false)
-                .isPreferredLeader(false)
-                .path("")
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(0)
-                .nodeInfo(NodeInfo.of(1, "broker0", 1111))
-                .lag(-1)
-                .size(-1)
-                .isLeader(false)
-                .isSync(true)
-                .isFuture(false)
-                .isOffline(false)
-                .isPreferredLeader(false)
-                .path("")
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(1)
-                .nodeInfo(NodeInfo.of(0, "broker0", 1111))
-                .lag(-1)
-                .size(-1)
-                .isLeader(true)
-                .isSync(true)
-                .isFuture(false)
-                .isOffline(false)
-                .isPreferredLeader(false)
-                .path("")
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(1)
-                .nodeInfo(NodeInfo.of(1, "broker0", 1111))
-                .lag(-1)
-                .size(-1)
-                .isLeader(false)
-                .isSync(true)
-                .isFuture(false)
-                .isOffline(false)
-                .isPreferredLeader(false)
-                .path("")
-                .build());
-    var after =
-        List.of(
-            Replica.builder()
-                .topic("topic1")
-                .partition(0)
-                .nodeInfo(NodeInfo.of(2, "broker0", 1111))
-                .lag(-1)
-                .size(-1)
-                .isLeader(true)
-                .isSync(true)
-                .isFuture(false)
-                .isOffline(false)
-                .isPreferredLeader(false)
-                .path("")
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(0)
-                .nodeInfo(NodeInfo.of(1, "broker0", 1111))
-                .lag(-1)
-                .size(-1)
-                .isLeader(false)
-                .isSync(true)
-                .isFuture(false)
-                .isOffline(false)
-                .isPreferredLeader(false)
-                .path("")
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(1)
-                .nodeInfo(NodeInfo.of(0, "broker0", 1111))
-                .lag(-1)
-                .size(-1)
-                .isLeader(true)
-                .isSync(true)
-                .isFuture(false)
-                .isOffline(false)
-                .isPreferredLeader(false)
-                .path("")
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(1)
-                .nodeInfo(NodeInfo.of(2, "broker0", 1111))
-                .lag(-1)
-                .size(-1)
-                .isLeader(false)
-                .isSync(true)
-                .isFuture(false)
-                .isOffline(false)
-                .isPreferredLeader(false)
-                .path("")
-                .build());
-    var beforeClusterInfo = ClusterInfoTest.of(before);
-    var afterClusterInfo = ClusterInfoTest.of(after);
-    var changedReplicaLeaderCount =
-        ClusterInfo.changedReplicaNumber(beforeClusterInfo, afterClusterInfo, Replica::isLeader);
-    Assertions.assertEquals(3, changedReplicaLeaderCount.size());
-    Assertions.assertTrue(changedReplicaLeaderCount.containsKey(0));
-    Assertions.assertTrue(changedReplicaLeaderCount.containsKey(1));
-    Assertions.assertTrue(changedReplicaLeaderCount.containsKey(2));
-    Assertions.assertEquals(-1, changedReplicaLeaderCount.get(0));
-    Assertions.assertEquals(0, changedReplicaLeaderCount.get(1));
-    Assertions.assertEquals(1, changedReplicaLeaderCount.get(2));
-  }
-
-  @Test
-  void testChangedReplicaNumber() {
-    var costFunction = new ReplicaNumberCost();
-    var before =
-        List.of(
-            Replica.builder()
-                .topic("topic1")
-                .partition(0)
-                .nodeInfo(NodeInfo.of(0, "broker0", 1111))
-                .lag(-1)
-                .size(-1)
-                .isLeader(true)
-                .isSync(true)
-                .isFuture(false)
-                .isOffline(false)
-                .isPreferredLeader(false)
-                .path("")
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(0)
-                .nodeInfo(NodeInfo.of(1, "broker0", 1111))
-                .lag(-1)
-                .size(-1)
-                .isLeader(false)
-                .isSync(true)
-                .isFuture(false)
-                .isOffline(false)
-                .isPreferredLeader(false)
-                .path("")
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(1)
-                .nodeInfo(NodeInfo.of(0, "broker0", 1111))
-                .lag(-1)
-                .size(-1)
-                .isLeader(true)
-                .isSync(true)
-                .isFuture(false)
-                .isOffline(false)
-                .isPreferredLeader(false)
-                .path("")
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(1)
-                .nodeInfo(NodeInfo.of(1, "broker0", 1111))
-                .lag(-1)
-                .size(-1)
-                .isLeader(false)
-                .isSync(true)
-                .isFuture(false)
-                .isOffline(false)
-                .isPreferredLeader(false)
-                .path("")
-                .build());
-    var after =
-        List.of(
-            Replica.builder()
-                .topic("topic1")
-                .partition(0)
-                .nodeInfo(NodeInfo.of(2, "broker0", 1111))
-                .lag(-1)
-                .size(-1)
-                .isLeader(true)
-                .isSync(true)
-                .isFuture(false)
-                .isOffline(false)
-                .isPreferredLeader(false)
-                .path("")
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(0)
-                .nodeInfo(NodeInfo.of(1, "broker0", 1111))
-                .lag(-1)
-                .size(-1)
-                .isLeader(false)
-                .isSync(true)
-                .isFuture(false)
-                .isOffline(false)
-                .isPreferredLeader(false)
-                .path("")
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(1)
-                .nodeInfo(NodeInfo.of(0, "broker0", 1111))
-                .lag(-1)
-                .size(-1)
-                .isLeader(true)
-                .isSync(true)
-                .isFuture(false)
-                .isOffline(false)
-                .isPreferredLeader(false)
-                .path("")
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(1)
-                .nodeInfo(NodeInfo.of(2, "broker0", 1111))
-                .lag(-1)
-                .size(-1)
-                .isLeader(false)
-                .isSync(true)
-                .isFuture(false)
-                .isOffline(false)
-                .isPreferredLeader(false)
-                .path("")
-                .build());
-    var beforeClusterInfo = ClusterInfoTest.of(before);
-    var afterClusterInfo = ClusterInfoTest.of(after);
-    var changedReplicaCount =
-        ClusterInfo.changedReplicaNumber(beforeClusterInfo, afterClusterInfo, ignored -> true);
-    Assertions.assertEquals(3, changedReplicaCount.size(), changedReplicaCount.toString());
-    Assertions.assertTrue(changedReplicaCount.containsKey(0));
-    Assertions.assertTrue(changedReplicaCount.containsKey(1));
-    Assertions.assertTrue(changedReplicaCount.containsKey(2));
-    Assertions.assertEquals(-1, changedReplicaCount.get(0));
-    Assertions.assertEquals(-1, changedReplicaCount.get(1));
-    Assertions.assertEquals(2, changedReplicaCount.get(2));
-  }
-
-  @Test
-  void testChangedRecordSizeOverflow() {
-    var limit = 1600;
-    var moveInResult = ClusterInfo.recordSizeToSync(beforeClusterInfo(), afterClusterInfo());
-    Assertions.assertEquals(3, moveInResult.size());
-    Assertions.assertEquals(0, moveInResult.get(0).bytes());
-    Assertions.assertEquals(1000, moveInResult.get(1).bytes());
-    Assertions.assertEquals(100 + 500, moveInResult.get(2).bytes());
-
-    var moveOutResult = ClusterInfo.recordSizeToFetch(beforeClusterInfo(), afterClusterInfo());
-    Assertions.assertEquals(3, moveOutResult.size());
-    Assertions.assertEquals(100 + 500, moveOutResult.get(0).bytes());
-    Assertions.assertEquals(0, moveOutResult.get(1).bytes());
-    Assertions.assertEquals(1000, moveOutResult.get(2).bytes());
-
-    var totalResult =
-        ClusterInfo.changedRecordSizeOverflow(
-            beforeClusterInfo(), afterClusterInfo(), ignored -> true, limit);
-    var overflowResult =
-        ClusterInfo.changedRecordSizeOverflow(
-            beforeClusterInfo(), afterClusterInfo(), ignored -> true, limit - 100);
-    Assertions.assertFalse(totalResult);
-    Assertions.assertTrue(overflowResult);
-  }
-
-  /*
-  before distribution:
-      p0: 0,1
-      p1: 0,1
-      p2: 2,0
-  after distribution:
-      p0: 2,1
-      p1: 0,2
-      p2: 1,0
-  leader log size:
-      p0: 100
-      p1: 500
-      p2  1000
-   */
-  private static ClusterInfo beforeClusterInfo() {
-    return ClusterInfo.of(
-        "fake",
-        List.of(NodeInfo.of(0, "aa", 22), NodeInfo.of(1, "aa", 22), NodeInfo.of(2, "aa", 22)),
-        Map.of(),
-        List.of(
-            Replica.builder()
-                .topic("topic1")
-                .partition(0)
-                .nodeInfo(NodeInfo.of(0, "broker0", 1111))
-                .size(100)
-                .isLeader(true)
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(0)
-                .nodeInfo(NodeInfo.of(1, "broker0", 1111))
-                .size(99)
-                .isLeader(false)
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(1)
-                .nodeInfo(NodeInfo.of(0, "broker0", 1111))
-                .size(500)
-                .isLeader(true)
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(1)
-                .nodeInfo(NodeInfo.of(1, "broker0", 1111))
-                .size(499)
-                .isLeader(false)
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(2)
-                .nodeInfo(NodeInfo.of(2, "broker0", 1111))
-                .size(1000)
-                .isLeader(true)
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(2)
-                .nodeInfo(NodeInfo.of(0, "broker0", 1111))
-                .size(1000)
-                .isLeader(false)
-                .build()));
-  }
-
-  private static ClusterInfo afterClusterInfo() {
-    return ClusterInfo.of(
-        "fake",
-        List.of(NodeInfo.of(0, "aa", 22), NodeInfo.of(1, "aa", 22), NodeInfo.of(2, "aa", 22)),
-        Map.of(),
-        List.of(
-            Replica.builder()
-                .topic("topic1")
-                .partition(0)
-                .nodeInfo(NodeInfo.of(2, "broker0", 1111))
-                .size(100)
-                .isLeader(true)
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(0)
-                .nodeInfo(NodeInfo.of(1, "broker0", 1111))
-                .size(99)
-                .isLeader(false)
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(1)
-                .nodeInfo(NodeInfo.of(0, "broker0", 1111))
-                .size(500)
-                .isLeader(true)
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(1)
-                .nodeInfo(NodeInfo.of(2, "broker0", 1111))
-                .size(500)
-                .isLeader(false)
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(2)
-                .nodeInfo(NodeInfo.of(1, "broker0", 1111))
-                .size(1000)
-                .isLeader(true)
-                .build(),
-            Replica.builder()
-                .topic("topic1")
-                .partition(2)
-                .nodeInfo(NodeInfo.of(0, "broker0", 1111))
-                .size(1000)
-                .isLeader(false)
-                .build()));
   }
 }
