+++ gui/src/main/java/org/astraea/gui/tab/health/BalancerNode.java
@@ -25,8 +25,8 @@ import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
 import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.BiConsumer;
 import java.util.function.BiFunction;
+import java.util.function.Consumer;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import javafx.scene.Node;
@@ -42,6 +42,7 @@ import org.astraea.common.balancer.algorithms.GreedyBalancer;
 import org.astraea.common.balancer.executor.RebalancePlanExecutor;
 import org.astraea.common.cost.HasClusterCost;
 import org.astraea.common.cost.HasMoveCost;
+import org.astraea.common.cost.MigrationCost;
 import org.astraea.common.cost.ReplicaLeaderCost;
 import org.astraea.common.cost.ReplicaLeaderSizeCost;
 import org.astraea.common.cost.ReplicaNumberCost;
@@ -89,34 +90,22 @@ class BalancerNode {
   static List<Map<String, Object>> costResult(Balancer.Plan plan) {
     var map = new HashMap<Integer, LinkedHashMap<String, Object>>();
 
-    BiConsumer<String, Map<Integer, ?>> process =
-        (name, m) ->
-            m.forEach(
-                (id, count) ->
-                    map.computeIfAbsent(
-                            id,
-                            ignored -> {
-                              var r = new LinkedHashMap<String, Object>();
-                              r.put("id", id);
-                              return r;
-                            })
-                        .put(name, count));
+    Consumer<List<MigrationCost>> process =
+        (migrationCosts) ->
+            migrationCosts.forEach(
+                migrationCost ->
+                    migrationCost.brokerCosts.forEach(
+                        (id, count) ->
+                            map.computeIfAbsent(
+                                    id,
+                                    ignored -> {
+                                      var r = new LinkedHashMap<String, Object>();
+                                      r.put("id", id);
+                                      return r;
+                                    })
+                                .put(migrationCost.name, count)));
 
-    process.accept(
-        "changed replicas",
-        ClusterInfo.changedReplicaNumber(
-            plan.initialClusterInfo(), plan.proposal(), ignored -> true));
-    process.accept(
-        "changed leaders",
-        ClusterInfo.changedReplicaNumber(
-            plan.initialClusterInfo(), plan.proposal(), Replica::isLeader));
-    process.accept(
-        "record size to fetch",
-        ClusterInfo.recordSizeToSync(plan.initialClusterInfo(), plan.proposal()));
-    process.accept(
-        "record size to sync",
-        ClusterInfo.recordSizeToFetch(plan.initialClusterInfo(), plan.proposal()));
-
+    process.accept(MigrationCost.migrationCosts(plan.initialClusterInfo(), plan.proposal()));
     return List.copyOf(map.values());
   }
